+++ src/datasources/products/components/ProductsQueryEditor.tsx
@@ -161,7 +161,7 @@ export function ProductsQueryEditor({ query, onChange, onRunQuery, datasource }:
           </div>
         </VerticalGroup>
       </HorizontalGroup>
-      <FloatingError message={datasource.error} />
+      <FloatingError message={datasource.errorTitle} innerMessage={datasource.errorDescription} severity="warning" />
     </>
   );
 }